Conflicts:
	src/system/boot/platform/u-boot/arch/ppc/Jamfile

+++ headers/os/drivers/PCI.h
@@ -44,15 +44,15 @@ typedef struct pci_info {
 	uchar	reserved;				/* filler, for alignment */
 	union {
 		struct {
-			ulong	cardbus_cis;			/* CardBus CIS pointer */
+			uint32	cardbus_cis;			/* CardBus CIS pointer */
 			ushort	subsystem_id;			/* subsystem (add-in card) id */
 			ushort	subsystem_vendor_id;	/* subsystem (add-in card) vendor id */
-			ulong	rom_base;				/* rom base address, viewed from host */
-			ulong	rom_base_pci;			/* rom base addr, viewed from pci */
-			ulong	rom_size;				/* rom size */
-			ulong	base_registers[6];		/* base registers, viewed from host */
-			ulong	base_registers_pci[6];	/* base registers, viewed from pci */
-			ulong	base_register_sizes[6];	/* size of what base regs point to */
+			uint32	rom_base;				/* rom base address, viewed from host */
+			uint32	rom_base_pci;			/* rom base addr, viewed from pci */
+			uint32	rom_size;				/* rom size */
+			uint32	base_registers[6];		/* base registers, viewed from host */
+			uint32	base_registers_pci[6];	/* base registers, viewed from pci */
+			uint32	base_register_sizes[6];	/* size of what base regs point to */
 			uchar	base_register_flags[6];	/* flags from base address fields */
 			uchar	interrupt_line;			/* interrupt line */
 			uchar	interrupt_pin;			/* interrupt pin */
@@ -60,9 +60,9 @@ typedef struct pci_info {
 			uchar	max_latency;			/* how often PCI access needed */
 		} h0;
 		struct {
-			ulong	base_registers[2];		/* base registers, viewed from host */
-			ulong	base_registers_pci[2];	/* base registers, viewed from pci */
-			ulong	base_register_sizes[2];	/* size of what base regs point to */
+			uint32	base_registers[2];		/* base registers, viewed from host */
+			uint32	base_registers_pci[2];	/* base registers, viewed from pci */
+			uint32	base_register_sizes[2];	/* size of what base regs point to */
 			uchar	base_register_flags[2];	/* flags from base address fields */
 			uchar	primary_bus;
 			uchar	secondary_bus;
@@ -75,12 +75,12 @@ typedef struct pci_info {
 			ushort	memory_limit;
 			ushort  prefetchable_memory_base;
 			ushort  prefetchable_memory_limit;
-			ulong	prefetchable_memory_base_upper32;
-			ulong	prefetchable_memory_limit_upper32;
+			uint32	prefetchable_memory_base_upper32;
+			uint32	prefetchable_memory_limit_upper32;
 			ushort	io_base_upper16;
 			ushort	io_limit_upper16;
-			ulong	rom_base;				/* rom base address, viewed from host */
-			ulong	rom_base_pci;			/* rom base addr, viewed from pci */
+			uint32	rom_base;				/* rom base address, viewed from host */
+			uint32	rom_base_pci;			/* rom base addr, viewed from pci */
 			uchar	interrupt_line;			/* interrupt line */
 			uchar	interrupt_pin;			/* interrupt pin */
 			ushort	bridge_control;
@@ -98,14 +98,14 @@ typedef struct pci_info {
 			uchar   subordinate_bus;
 			uchar   secondary_latency;
 			ushort  reserved;
-			ulong   memory_base;
-			ulong   memory_limit;
-			ulong   memory_base_upper32;
-			ulong   memory_limit_upper32;
-			ulong   io_base;
-			ulong   io_limit;
-			ulong   io_base_upper32;
-			ulong   io_limit_upper32;
+			uint32  memory_base;
+			uint32  memory_limit;
+			uint32  memory_base_upper32;
+			uint32  memory_limit_upper32;
+			uint32  io_base;
+			uint32  io_limit;
+			uint32  io_base_upper32;
+			uint32  io_limit_upper32;
 			ushort  secondary_status;
 			ushort  bridge_control;
 #endif /* __HAIKU_PCI_BUS_MANAGER_TESTING */
